# ARROW-8844: [C++] Transfer bitmap in words