Copy/Invert bitmap in words improves performance significantly.
Partial aligned case jumps from 914M/s to 6.6G/s, and non aligned
case from 82M/s to 3.8G/s on my test machine.