@cyb70289 2020-05-29
Tested on AMD Zen1 with gcc-7.5

before (CopyBitmapWithOffsetBoth is added by this patch)
```
CopyBitmapWithoutOffset/8192                  190 ns          190 ns      3678223 bytes_per_second=40.0804G/s
CopyBitmapWithOffset/8192                    8548 ns         8548 ns        81846 bytes_per_second=914.007M/s
CopyBitmapWithOffsetBoth/8192               95447 ns        95445 ns         7334 bytes_per_second=81.8531M/s
```

after
```
CopyBitmapWithoutOffset/8192                  191 ns          191 ns      3670699 bytes_per_second=39.9939G/s
CopyBitmapWithOffset/8192                    1151 ns         1151 ns       600632 bytes_per_second=6.62768G/s
CopyBitmapWithOffsetBoth/8192                2018 ns         2018 ns       346761 bytes_per_second=3.78089G/s
```