@cyb70289 2020-05-29
There are similar codes in my three recent patches optimizing unaligned bitmap operations (logical, compare, copy). I opened a new jira to refine. https://issues.apache.org/jira/browse/ARROW-8979