@fsaintjacques 2020-05-29
> [AMD64 Ubuntu 18.04 C++ Benchmark (#108408)](https://ci.ursalabs.org/#builders/73/builds/65) builder has been succeeded.
> 
> Revision: [ef4d77c](https://github.com/apache/arrow/commit/ef4d77ceabd23b51e052e4da58310ebacdc0fe9d)
> 
> ```diff
>   =============================  ===========  ===========  ============
>   benchmark                         baseline    contender        change
>   =============================  ===========  ===========  ============
>   CopyBitmapWithOffsetBoth/8192  2.48133e+09  2.48132e+09  -3.12173e-06
>   CopyBitmapWithOffset/8192      4.14268e+09  4.14273e+09   1.20158e-05
> ```

I'll need to fix this, tried to address this last time, weirdly it was fixed locally...