@wesm 2020-05-31
Is there a JIRA about fixing the legibility of the benchmark diff?