@fsaintjacques 2020-06-01
@ursabot benchmark --suite-filter=arrow-bit-util-benchmark --benchmark-filter=CopyBitmapWithOffset origin/master