@fsaintjacques 2020-06-01
once https://github.com/ursa-labs/ursabot/pull/197 is merged, we shouldn't need to specify `origin/master` as a baseline.